+++ templates/main.php
@@ -16,6 +16,7 @@ Util::addStyle('analytics', 'dashboard');
 Util::addStyle('analytics', 'wizard');
 Util::addStyle('analytics', 'print');
 Util::addScript('analytics', 'app');
+Util::addScript('analytics', 'chartOptions');
 Util::addScript('analytics', 'visualization');
 Util::addScript('analytics', 'sidebar');
 Util::addScript('analytics', 'navigation');